Extract observer tests
